refactor(#258): extract shared Langfuse helpers, remove unused flag, tag error paths
- New analytics/query_engine/langfuse_utils.py centralises the optional-
  import shim (lf_observe, lf_context) and report_langfuse_usage() so the
  three Q&A modules no longer duplicate ~20 lines of identical logic.
- Remove unused _HAS_LANGFUSE variable from intent.py, routing.py, and
  synthesis.py (was assigned but never read).
- Tag failed classification paths (llm_failed, invalid_json,
  schema_validation) with WARNING-level Langfuse observations so they
  surface in the trace UI alongside the existing ERROR tag for exceptions.

Made-with: Cursor
