+++ analytics/query_engine/intent.py
@@ -22,36 +22,15 @@ from typing import TYPE_CHECKING, Any, Final
 import structlog
 from pydantic import BaseModel, Field, field_validator
 
+from analytics.query_engine.langfuse_utils import lf_context as langfuse_context
+from analytics.query_engine.langfuse_utils import lf_observe as _lf_observe
+from analytics.query_engine.langfuse_utils import report_langfuse_usage
 from common.llm_adapter import complete
 
 if TYPE_CHECKING:
     from analytics.query_engine.schemas import CostLedger
 
 log = structlog.get_logger()
-
-# ---------------------------------------------------------------------------
-# Langfuse @observe — optional; no-op when SDK is absent or unconfigured
-# ---------------------------------------------------------------------------
-try:
-    from langfuse.decorators import langfuse_context
-    from langfuse.decorators import observe as _lf_observe
-
-    _HAS_LANGFUSE = True
-except ImportError:
-    _HAS_LANGFUSE = False
-
-    def _lf_observe(**_kwargs: Any):  # type: ignore[misc]
-        def _decorator(fn: Any) -> Any:
-            return fn
-
-        return _decorator
-
-    class _FakeLangfuseContext:
-        @staticmethod
-        def update_current_observation(**_kwargs: Any) -> None:
-            pass
-
-    langfuse_context = _FakeLangfuseContext()  # type: ignore[assignment]
 
 _AGENT_NAME = "analytics-intent-classification"
 _CLARIFICATION_THRESHOLD = 0.55
@@ -243,27 +222,6 @@ def _fallback_other(reason: str) -> dict[str, Any]:
     }
 
 
-def _update_langfuse_usage(llm_result: dict[str, Any]) -> None:
-    """Propagate token counts and model name from a complete() result to the current Langfuse observation."""
-    input_tokens = llm_result.get("input_tokens") or llm_result.get("prompt_tokens")
-    output_tokens = llm_result.get("output_tokens") or llm_result.get("completion_tokens")
-    model = llm_result.get("model")
-    cost_usd = llm_result.get("cost_usd")
-    update_kwargs: dict[str, Any] = {}
-    if input_tokens is not None or output_tokens is not None:
-        update_kwargs["usage"] = {
-            "input": int(input_tokens or 0),
-            "output": int(output_tokens or 0),
-            "total": int((input_tokens or 0) + (output_tokens or 0)),
-        }
-    if model:
-        update_kwargs["model"] = str(model)
-    if cost_usd is not None:
-        update_kwargs["cost_details"] = {"total": float(cost_usd)}
-    if update_kwargs:
-        langfuse_context.update_current_observation(**update_kwargs)
-
-
 @_lf_observe(as_type="generation", name="intent_classification")
 def classify_workforce_question(
     question: str,
@@ -322,19 +280,22 @@ def classify_workforce_question(
         langfuse_context.update_current_observation(level="ERROR", status_message=str(exc))
         return _fallback_other("llm_exception")
 
-    _update_langfuse_usage(result)
+    report_langfuse_usage(result)
 
     if not result.get("success") or result.get("extraction_failed"):
+        langfuse_context.update_current_observation(level="WARNING", status_message="llm_failed")
         return _fallback_other("llm_failed")
 
     content = (result.get("content") or "").strip()
     parsed = _parse_llm_json(content)
     if not parsed:
+        langfuse_context.update_current_observation(level="WARNING", status_message="invalid_json")
         return _fallback_other("invalid_json")
 
     try:
         validated = IntentClassification.model_validate(parsed)
     except Exception:
+        langfuse_context.update_current_observation(level="WARNING", status_message="schema_validation")
         return _fallback_other("schema_validation")
 
     out_entities = validated.extracted_entities.model_dump()
